Fix mad-eye/integration-tests#20.
Added files that have a newer mtime are marked
as modified and modified_locally.

We wanted to also reset bolide's contents, but
that's tricky, since the model object doesn't have
access to bolideClient, and it can't remove the document from
sharejs anyway.  It could refresh it, which would be a next step.